Adding preset color management - END

+++ src/app/admin/api/theme/preset-color/delete/[id]/route.ts
@@ -3,6 +3,8 @@ import { getServerSession } from "next-auth/next";
 import prisma from "@/lib/db/prisma";
 import { authOptions } from "@/utils/authOptions";
 import { NextResponse } from "next/server";
+import { OnlyString } from "@/app/api/theme/theme";
+import { Theme } from "@prisma/client";
 
 export async function GET(
   req: Request,
@@ -13,10 +15,49 @@ export async function GET(
   if (session) {
     try {
       const id = parseInt(params.id);
-      await prisma.presetColor.delete({
-        where: { id },
+      const presetColor = await prisma.presetColor.findUnique({
+        where: {
+          id,
+        },
       });
-      return NextResponse.json({ message: "ok" }, { status: 200 });
+
+      if (presetColor) {
+        const themes = await prisma.theme.findMany();
+        const themesToUpdate: Theme[] = [];
+
+        themes.forEach((t) => {
+          Object.entries(t).forEach(([key, value]) => {
+            if (typeof value === "string" && value === presetColor.name) {
+              t[key as keyof OnlyString<Theme>] = presetColor.color;
+              themesToUpdate.push(t);
+            }
+          });
+        });
+
+        for await (const theme of themesToUpdate) {
+          const { id, ...data } = theme;
+          await prisma.theme.update({
+            where: {
+              id,
+            },
+            data: {
+              ...data,
+            },
+          });
+        }
+
+        await prisma.presetColor.delete({
+          where: { id },
+        });
+
+        const updatedThemes = await prisma.theme.findMany();
+        const updatedPresetColors = await prisma.presetColor.findMany();
+
+        return NextResponse.json({
+          updatedThemes: JSON.parse(JSON.stringify(updatedThemes)),
+          updatedPresetColors: JSON.parse(JSON.stringify(updatedPresetColors)),
+        });
+      }
     } catch (e) {
       console.log(e);
       return NextResponse.json({ error: "Error" }, { status: 404 });
